FOC/L4RE: Upstream revision 72

+++ kernel/fiasco/src/jdb/ia32/jdb_io_ports.cpp
@@ -127,7 +127,7 @@ Io_m::action( int cmd, void *&args, char const *&fmt, int &)
 		  else
 		    {
 		      answer = Io::in8(buf.io.adr);
-		      printf("%02x\n", (int)answer);
+		      printf("%02x\n", answer);
 		    }
 		}
 	      else // out
