+++ kernel/fiasco/src/jdb/ia32/jdb_kern_info-pci.cpp
@@ -58,8 +58,9 @@ Jdb_kern_info_pci::show()
               if (classcode == 0x06 && subclass == 0x04)
                 buses++;
 
-              printf ("%02lx:%02lx.%1lx Class %02x%02x: %04x:%04x ",
-                  bus, dev, subdev, classcode, subclass, device, vendor);
+              printf("%02lx:%02lx.%1lx Class %02x%02x: %04x:%04x ",
+                     bus, dev, subdev, (unsigned)classcode,
+                     (unsigned)subclass, (unsigned)device, (unsigned)vendor);
               if (classcode < sizeof(classes)/sizeof(classes[0]))
                 printf("%s", classes[classcode]);
               putchar('\n');
